Cleaned up Geofence code.
