+++ PokeAlarm/Filters.py
@@ -460,42 +460,3 @@ class GymFilter(Filter):
                 sys.exit(1)
         return s
 
-
-class Geofence(object):
-
-    # Expects points to be
-    def __init__(self, name, points):
-        self.name = name
-        self.__points = points
-
-        self.__min_x = points[0][0]
-        self.__max_x = points[0][0]
-        self.__min_y = points[0][1]
-        self.__max_y = points[0][1]
-
-        for p in points:
-            self.__min_x = min(p[0], self.__min_x)
-            self.__max_x = max(p[0], self.__max_x)
-            self.__min_y = min(p[1], self.__min_y)
-            self.__max_y = max(p[1], self.__max_y)
-
-    def contains(self, x, y):
-        # Quick check the boundary box of the entire polygon
-        if self.__max_x < x or x < self.__min_x or self.__max_y < y or y < self.__min_y:
-            return False
-
-        inside = False
-        p1x, p1y = self.__points[0]
-        n = len(self.__points)
-        for i in range(1, n+1):
-            p2x, p2y = self.__points[i % n]
-            if min(p1y, p2y) < y <= max(p1y, p2y) and x <= max(p1x, p2x):
-                        if p1y != p2y:
-                            xinters = (y - p1y) * (p2x - p1x) / (p2y - p1y) + p1x
-                        if p1x == p2x or x <= xinters:
-                            inside = not inside
-            p1x, p1y = p2x, p2y
-        return inside
-
-    def get_name(self):
-        return self.__name
